+++ package.json
@@ -37,7 +37,6 @@
     "build": "react-scripts build",
     "build:client": "vite build --outDir dist/client",
     "build:server": "vite build --outDir dist/server --ssr src/enterServer.tsx",
-    "build:ssr": "npm run build:client && npm run build:server",
     "server": "npx ts-node --esm server.ts",
     "test": "react-scripts test --coverage --watchAll",
     "predeploy": "npm run build",
